added JpaSpecificationExecutor support for every repo

+++ rd-api/src/main/java/it/clevercom/echo/rd/repository/IOrderService_rd_Repository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.clevercom.echo.rd.model.entity.OrderService;
 
-public interface IOrderService_rd_Repository extends JpaRepository<OrderService, Integer> {
+public interface IOrderService_rd_Repository extends JpaRepository<OrderService, Long> {
 
 }
